+++ src/app/profile/profile.page.ts
@@ -33,7 +33,7 @@ export class ProfilePage implements OnInit {
     private toastController: ToastController,
     private swUpdate: SwUpdate) {
 
-    this.version = "7.1.0";
+    this.version = "7.2.0";
     this.user = new User('','');
   }