+++ src/app/race-result/race-result.page.ts
@@ -20,7 +20,7 @@ import { ActivatedRoute } from '@angular/router';
 export class RaceResultPage {
   drivers: Array<Driver> = Driver.all();
   result: Result = new Result("");
-  seasonResult: SeasonResult = new SeasonResult(2024);
+  seasonResult: SeasonResult = new SeasonResult(2025);
   driversOrdered: Array<Driver>;
   isAdmin: boolean = false;
   selectedRace: Race = Race.empty();